+++ src/main/java/swp391_gr7/hivsystem/dto/request/ConsultationCreateRequest.java
@@ -9,6 +9,6 @@ import java.time.LocalDate;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ConsultationCreateRequest {
-    private Long appointmentId;
+    private int appointmentId;
     private String notes;
 }
