+++ src/main/java/swp391_gr7/hivsystem/repository/OAuth2Repository.java
@@ -3,7 +3,7 @@ package swp391_gr7.hivsystem.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import swp391_gr7.hivsystem.model.Users;
 
-public interface OAuth2Repository extends JpaRepository<Users, Long> {
+public interface OAuth2Repository extends JpaRepository<Users, Integer> {
     boolean existsByEmail(String email);
     boolean existsByPhone(String phone);
     boolean existsByUsername(String username);
