+++ src/main/java/swp391_gr7/hivsystem/repository/TestResultRepository.java
@@ -7,10 +7,10 @@ import swp391_gr7.hivsystem.model.TestResults;
 
 import java.util.List;
 @Repository
-public interface TestResultRepository extends JpaRepository<TestResults, Long> {
-    List<TestResults> findByCustomers_CustomerId(Long customerId);
+public interface TestResultRepository extends JpaRepository<TestResults, Integer> {
+    List<TestResults> findByCustomers_CustomerId(int customerId);
 
-    List<TestResults> findByAppointments_AppointmentId(Long appointmentId);
+    List<TestResults> findByAppointments_AppointmentId(int appointmentId);
 
     TestResults save(TestResults testResults);
